+++ mayan/apps/dynamic_search/api_views.py
@@ -9,6 +9,7 @@ from mayan.apps.rest_api import generics
 from .classes import SearchModel
 from .mixins import SearchModelMixin
 from .serializers import SearchModelSerializer
+from .runtime import search_backend
 
 
 class APISearchView(SearchModelMixin, generics.ListAPIView):
@@ -26,7 +27,8 @@ class APISearchView(SearchModelMixin, generics.ListAPIView):
             self.mayan_object_permissions = {'GET': (search_model.permission,)}
 
         try:
-            queryset = search_model.search(
+            queryset = search_backend.search(
+                search_model=search_model,
                 query_string=self.request.GET, user=self.request.user
             )
         except Exception as exception:
@@ -63,9 +65,10 @@ class APIAdvancedSearchView(SearchModelMixin, generics.ListAPIView):
             global_and_search = False
 
         try:
-            queryset = self.search_model.search(
+            queryset = search_backend.search(
                 global_and_search=global_and_search,
-                query_string=self.request.GET, user=self.request.user
+                query_string=self.request.GET,
+                search_model=self.search_model, user=self.request.user
             )
         except Exception as exception:
             raise ParseError(force_text(exception))